+++ src/components/MatchCriterion/edit/ExistingRecordSections/ExistingSectionFolio.js
@@ -1,10 +1,8 @@
-import React, { useState } from 'react';
+import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
+import { Field } from 'redux-form';
 import classnames from 'classnames';
-import {
-  last,
-  noop,
-} from 'lodash';
+import { last } from 'lodash';
 
 import {
   Row,
@@ -20,6 +18,7 @@ import {
 import {
   fieldsConfig,
   MARC_FIELD_CONSTITUENT,
+  validateRequiredField,
 } from '../../../../utils';
 
 import css from '../MatchCriterions.css';
@@ -28,16 +27,15 @@ export const ExistingSectionFolio = ({
   repeatableIndex,
   existingRecordFieldLabel,
   existingRecordFields,
-  existingRecordFieldsValue,
   existingRecordType,
-  dispatchFormChange,
 }) => {
-  const [isDirty, setDirty] = useState(false);
+  const fieldToChangeName = `profile.matchDetails[${repeatableIndex}].existingMatchExpression.fields`;
 
-  const dropdownValue = last(existingRecordFieldsValue).value;
+  const handleFieldSearch = (value, dataOptions) => {
+    return dataOptions.filter(o => new RegExp(`${value}`, 'i').test(o.label));
+  };
 
   const handleExistingRecordSelect = value => {
-    const fieldToChangeName = `profile.matchDetails[${repeatableIndex}].existingMatchExpression.fields`;
     const fieldId = existingRecordFields.find(item => item.value === value)?.id;
     const fieldFromConfig = fieldsConfig.find(item => item.id === fieldId && item.recordType === existingRecordType);
     const fieldToChangeValue = [{
@@ -54,13 +52,27 @@ export const ExistingSectionFolio = ({
       });
     }
 
-    dispatchFormChange(fieldToChangeName, fieldToChangeValue);
-    setDirty(value !== dropdownValue);
+    return fieldToChangeValue;
   };
 
-  const handleFieldSearch = (value, dataOptions) => {
-    return dataOptions.filter(o => new RegExp(`${value}`, 'i').test(o.label));
-  };
+  const formatExistingRecordValue = useCallback(
+    value => {
+      const fieldValue = value[0].value;
+      const fieldFromConfig = fieldsConfig.find(item => item.value === fieldValue && item.recordType === existingRecordType);
+
+      if (fieldFromConfig?.fromResources) {
+        return last(value).value;
+      }
+
+      return existingRecordFields.find(item => item.id === fieldFromConfig?.id)?.value;
+    },
+    [existingRecordFields, existingRecordType],
+  );
+
+  const validateExistingFieldValue = useCallback(
+    value => validateRequiredField(last(value).value),
+    [],
+  );
 
   return (
     <Section
@@ -69,14 +81,17 @@ export const ExistingSectionFolio = ({
     >
       <Row>
         <Col xs={12}>
-          <Selection
+          <Field
             id="criterion-value-type"
-            value={dropdownValue}
-            dataOptions={existingRecordFields}
-            onChange={handleExistingRecordSelect}
-            onFilter={handleFieldSearch}
-            dirty={isDirty}
             aria-label={existingRecordFieldLabel}
+            component={Selection}
+            name={fieldToChangeName}
+            dataOptions={existingRecordFields}
+            format={formatExistingRecordValue}
+            parse={handleExistingRecordSelect}
+            onFilter={handleFieldSearch}
+            onBlur={e => e.preventDefault()}
+            validate={[validateExistingFieldValue]}
           />
         </Col>
       </Row>
@@ -86,18 +101,15 @@ export const ExistingSectionFolio = ({
 
 ExistingSectionFolio.propTypes = {
   repeatableIndex: PropTypes.number.isRequired,
-  existingRecordFieldsValue: PropTypes.arrayOf(PropTypes.object).isRequired,
   existingRecordType: PropTypes.oneOf(Object.keys(FOLIO_RECORD_TYPES)).isRequired,
   existingRecordFields: PropTypes.arrayOf(PropTypes.shape({
     value: PropTypes.string.isRequired,
     label: PropTypes.string.isRequired,
   })),
   existingRecordFieldLabel: PropTypes.string,
-  dispatchFormChange: PropTypes.func,
 };
 
 ExistingSectionFolio.defaultProps = {
   existingRecordFields: null,
   existingRecordFieldLabel: '',
-  dispatchFormChange: noop,
 };